Working code for improved ed
