+++ .direnv/flake-profile
@@ -1,1 +1,1 @@
-flake-profile-5-link
+flake-profile-6-link
