+++ .direnv/flake-profile-6-link
@@ -1,0 +1,1 @@
+/nix/store/57a6dfgplvcn4j2k2cwb5r1yhrzvhb15-interactive-python3-3.9.9-environment-env
