--- ising_energy_square_9.json
@@ -1,12 +1,0 @@
-{
-	"nlce_type": "square",
-	"final_order": 9,
-	"model": "ising",
-	"property_type": "energy",
-	"temp_range": [0.5, 20],
-	"granularity": 500,
-	"starting_order_plot": 6,
-	"nlce_data_dir": "./Data/NLCE_Data",
-	"output_dir": "./Data/Property_Data",
-	"benchmarking": false
-}
